+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- dtoolcore.utils.get_config_value_from_file
+
 
 Changed
 ^^^^^^^